requirements.txt
asgiref==3.8.1
Django==5.0.5
django-cors-headers==4.3.1
djangorestframework==3.15.1
Jinja2==3.1.4
MarkupSafe==2.1.5
ordered-set==4.1.0
pillow==10.3.0
PyLaTeX==1.4.2
python-dotenv==1.0.1
sqlparse==0.5.0
